Redo both Cryptostream internals
This allows small reads on bufread::Cryptostream, which would not have
enough space to store at least one more block, as required by OpenSSL.
When one calls BufRead::read_to_end, it starts out with a 32 byte read,
which classifies as a small read.

It also allows big writes on write::Cryptostream, which won't work if
the internal buffer is smaller than the amount of data written plus that
one additional block size. Big writes are the default when one does a
std::io::copy with write::Encryptor or write::Decryptor as the
destination.

Also the commented eprintln! statements have been replaced with trace!
so a recompile isn't needed in order to debug issues.

Last, but not least, the hardcoded vendored feature for openssl is
replaced with an openssl-vendored feature in this crate so one isn't
forced to recompile openssl on all platforms.