Simplify buffering and fix handling of zero-length streams
The design now embraces the OpenSSL bindings used for encryption rather
than assuming that writes of partial blocks will not be possible. As a
result, it is OK to execute individual writes to a cryptostream smaller
than a block size without fear of the stream being finalized.

Also addresses neosmart/cryptostream#5 for both reads and writes
(for both encryptions and decryptions) by allowing for zero-length
streams to skip finalization (which triggers a panic in the OpenSSL
bindings).