Add option to use system OpenSSL
By default, we use the openssl dependency with the `vendored` feature,
which causes it to be built and statically linked at runtime.

The new cryptostream crate's `system-openssl` unsets this option and
causes the library to depend on the system libssl.

If there is a cleaner way to do this without such an ugly hack, please
let me know!