Fix incorrect read offset with extremely small buffer sizes
In the case where an initial go around the update loop caused bytes to
be read from the source stream and fed into the crypter, but the crypter
did not return any bytes (i.e. less than a block of untransformed data
was fed into the crypter), if the second time around the loop we ended
up reading enough bytes in aggregate to cause the crypter to output a
block but did not require the use of the write buffer, we would end up
re-transforming the same previously read bytes.

Closes #7